Allow lookup of MainModuleName to fail (#99)
I don't fully understand in which situations this may fail.
MSDN claims that a Win32Exception is raised when a 32-bit process
attempts to access the modules of a 64-bit process.
Internally, looking the main module of a process results in a call to
the OpenProcess win32 API function, which suggests that the operation
may also fail on certain system-level processes when user-level code
attempts to access them.

In any case, this is a non-critical operation. Cleaning up
non-responsive GPG-agent processes is merely a convenience feature
and should be done on a best-effort basis only, since it's not really
our job to ensure that GPG remains accessible.