Disable 'save' button unless file is modified
This is targeting the case when a user types or generates the password in
generator's textbox thinking it will be saved to the file. The user
should copy the password to the 1st line of the file content textxbox
before saving but they might forget about it. Disabling the button
prevents resaving the same content to the file, creating a dumb git
commit and loosing the pasword (if the user doesn't have a copy of it).